chore(test): split unit and integration test lanes
Move 26 test files into tests/unit/ (18) and tests/integration/ (8),
keeping shared helpers (helpers.mjs, install-paths.mjs, mocks/) at
tests/ root. Add `test:unit` and `test:int` scripts; `test` still runs
the full suite for backwards compatibility.

Integration lane = tests that spawn subprocesses or do heavy fs/git
work (install setup, broker lifecycle, gemini-companion CLI, etc.).
Unit lane = everything else.

All 203 tests still pass. Fast feedback loop drops to 14.2s
(unit-only) from 19.1s (full suite).
