+++ tests/integration/broker-lifecycle.test.mjs
@@ -30,7 +30,7 @@ import {
   BROKER_PID_FILENAME,
   BROKER_LOG_FILENAME,
   BROKER_SESSION_FILENAME
-} from "./install-paths.mjs";
+} from "../install-paths.mjs";
 
 const LIB_DIR = path.join(PLUGIN_ROOT, PLUGIN_SOURCE_DIR_RELATIVE, "scripts", "lib");
