+++ tests/integration/gemini-companion-flags.test.mjs
@@ -4,7 +4,7 @@ import test from "node:test";
 import assert from "node:assert/strict";
 import { fileURLToPath } from "node:url";
 
-const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
+const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..", "..");
 const COMPANION_SRC = fs.readFileSync(path.join(ROOT, "plugins/gemini/scripts/gemini-companion.mjs"), "utf8");
 
 function functionSource(name) {
